+++ k2/src/main/kotlin/at/asitplus/gradle/CommonCompilerOptions.kt
@@ -1,11 +1,9 @@
 @file:Suppress("NOTHING_TO_INLINE")
+
 package at.asitplus.gradle
 
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import org.jetbrains.kotlin.gradle.kpm.external.ExternalVariantApi
-import org.jetbrains.kotlin.gradle.kpm.external.project
 
-@OptIn(ExternalVariantApi::class)
 internal inline fun KotlinMultiplatformExtension.experimentalOptIns() {
     Logger.lifecycle("  Adding opt ins")
     Logger.info("   * Serialization")